% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{extract_validation_errors}
\alias{extract_validation_errors}
\title{Extract errors from validation results}
\usage{
extract_validation_errors(results)
}
\arguments{
\item{results}{Validation results object from \code{validate::confront()}}
}
\value{
A tibble with columns: rule, error
}
\description{
Processes validation results to extract any errors thrown by validator
}
