% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_helpers.R
\name{is_valid_code_list}
\alias{is_valid_code_list}
\title{Check if string is valid code list}
\usage{
is_valid_code_list(value, codes)
}
\arguments{
\item{value}{Character string to validate as comma-separates list of values}

\item{codes}{Character vector of valid code values}
}
\value{
Logical value: TRUE if valid code list or empty/NA, FALSE otherwise
}
\description{
Validates whether a string contains a comma-separated
list of valid codes from a predefined list.
Empty strings and NA values are considered valid.
}
\examples{
is_valid_code_list('x',    c('x', 'y')) # TRUE
is_valid_code_list('x,y',  c('x', 'y')) # TRUE
is_valid_code_list('x, y', c('x', 'y')) # TRUE
is_valid_code_list('z',    c('x', 'y')) # FALSE
is_valid_code_list('x,z',  c('x', 'y')) # FALSE
is_valid_code_list(NA, c('x', 'y'))     # TRUE
is_valid_code_list("", c('x', 'y'))     # TRUE

}
