% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makaraValidatr-package.R
\docType{package}
\name{makaraValidatr-package}
\alias{makaraValidatr}
\alias{makaraValidatr-package}
\title{makaraValidatr: Data validation tools for Makara passive acoustics monitoring database}
\description{
Tools for validating data in the Makara passive acoustics monitoring database. Includes functions for checking data types, ranges, and consistency. Designed for validating template files prior to submission to the Makara database.
}
\author{
\strong{Maintainer}: Jeff Walker \email{jeffrey.walker@noaa.gov}

}
\keyword{internal}
