% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_DATE}
\alias{MIN_DATE}
\title{Minimum valid date}
\format{
A character string in ISO 8601 format (YYYY-MM-DD)
}
\usage{
MIN_DATE
}
\description{
The earliest date considered valid for date fields in Makara database.
}
\keyword{datasets}
