% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepare_data_code_list}
\alias{prepare_data_code_list}
\title{Add code list validation columns}
\usage{
prepare_data_code_list(data, columns, reference_tables = list())
}
\arguments{
\item{data}{A data frame to process}

\item{columns}{A data frame of column definitions containing code list info
with \code{name}, \code{reference_table} and \code{transaction_table} columns}

\item{reference_tables}{A named list of reference tables, each containing a
data frame with a \code{code} column}
}
\value{
Data frame with additional code list validation columns
}
\description{
For each code list column in the data, adds a corresponding boolean column
indicating whether the list contains valied codes. The new column is named
\code{<column_name>_valid_code_list}.
}
\examples{
\dontrun{
data <- data.frame(tab_code = c("x,y", "z"))
prepare_data_code_list(
  data,
  list(name = "tab_code", reference_table = "tab"),
  list(tab = c("x", "y"))
)
}

}
